refactor: JpaRepository를 infra - persistence 모듈로 이동
