+++ src/main/java/kitchenpos/products/infra/persistence/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.infra.persistence;
 
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
